+++ CHANGELOG.md
@@ -9,7 +9,7 @@
 - https://github.com/creativetimofficial/ct-nextjs-material-kit-pro/issues/2
 ### Major style changes
 - `assets/scss/plugins/_plugin-react-image-gallery.scss`
-- `assets/jss/nextjs-material-kit-pro/pages/productStyle.js`
+- `assets/jss/nextjs-material-kit-pro/pages/examples/productStyle.js`
 - `assets/jss/nextjs-material-kit-pro/components/customDropdownStyle.js`
 ### Deleted components
 ### Added components